client/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
numpy==1.26.3
opencv-python==4.9.0.80
pillow==10.2.0
requests==2.31.0
urllib3==2.1.0
